+++ src/Exceptions/NonRetryableException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace YdbPlatform\Ydb\Exceptions;
+
+class NonRetryableException extends \Exception
+{
+}