+++ src/Exceptions/Ydb/AlreadyExistsException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace YdbPlatform\Ydb\Exceptions\Ydb;
+
+class AlreadyExistsException extends \YdbPlatform\Ydb\Exceptions\NonRetryableException
+{
+
+}