+++ src/Exceptions/Ydb/BadRequestException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace YdbPlatform\Ydb\Exceptions\Ydb;
+
+use YdbPlatform\Ydb\Exceptions\NonRetryableException;
+
+class BadRequestException extends NonRetryableException
+{
+
+}