+++ src/Exceptions/Ydb/ClientResourceExhaustedException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace YdbPlatform\Ydb\Exceptions\Ydb;
+
+class ClientResourceExhaustedException extends \YdbPlatform\Ydb\Exceptions\NonRetryableException
+{
+
+}