+++ src/Exceptions/Ydb/NotFoundException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace YdbPlatform\Ydb\Exceptions\Ydb;
+
+class NotFoundException extends \YdbPlatform\Ydb\Exceptions\NonRetryableException
+{
+
+}