+++ src/Exceptions/Ydb/OverloadException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace YdbPlatform\Ydb\Exceptions\Ydb;
+
+class OverloadException extends \YdbPlatform\Ydb\Exceptions\RetryableException
+{
+    public function isFastBackoff(): bool
+    {
+        return false;
+    }
+}