+++ src/Exceptions/Ydb/PreconditionFailedException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace YdbPlatform\Ydb\Exceptions\Ydb;
+
+class PreconditionFailedException extends \YdbPlatform\Ydb\Exceptions\NonRetryableException
+{
+
+}