+++ src/Exceptions/Ydb/SchemeError.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace YdbPlatform\Ydb\Exceptions\Ydb;
+
+class SchemeError extends \YdbPlatform\Ydb\Exceptions\NonRetryableException
+{
+
+}