+++ src/Exceptions/Ydb/SessionExpiredException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace YdbPlatform\Ydb\Exceptions\Ydb;
+
+class SessionExpiredException extends \YdbPlatform\Ydb\Exceptions\NonRetryableException
+{
+
+}