+++ src/Exceptions/Ydb/TimeoutException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace YdbPlatform\Ydb\Exceptions\Ydb;
+
+class TimeoutException extends \YdbPlatform\Ydb\Exceptions\NonRetryableException
+{
+
+}