+++ src/Exceptions/Ydb/UnavailableException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace YdbPlatform\Ydb\Exceptions\Ydb;
+
+class UnavailableException extends \YdbPlatform\Ydb\Exceptions\RetryableException
+{
+    public function isFastBackoff(): bool
+    {
+        return true;
+    }
+
+}