+++ src/Exceptions/Ydb/UndeterminedException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace YdbPlatform\Ydb\Exceptions\Ydb;
+
+use YdbPlatform\Ydb\Exceptions\RetryableException;
+
+class UndeterminedException extends RetryableException
+{
+    public function isFastBackoff(): bool
+    {
+        return true;
+    }
+}