+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* added retry function
+
 ## 1.5.5
 
 * fixed iam auth for PHP < 8.0